underwriting_assistant/requirements.txt
flask
langchain
langchain-openai
python-dotenv
